sheratan-journal/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
jsonschema>=4.23.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
PyPDF2>=3.0.1
pdfplumber>=0.11.4
